Random Validation
Method created for validating a word against the part of speech.
Added this validation to words the user types in (checks if a typed
word is a noun if a noun is asked for) and now a user can request
a random word that matches the part of speech being asked for.